requirements.txt
streamlit
pandas
numpy
matplotlib
requests
beautifulsoup4
lxml
tabulate
pyyaml
google-generativeai>=0.7.2
kiteconnect>=4.2.0
# If available for your account setup:
growwapi>=0.1.0
